+++ src/Elba/Graphics/GraphicsModule.cpp
@@ -72,8 +72,8 @@ bool GraphicsModule::DeregisterForResize(GlobalKey key)
 
 void GraphicsModule::OnResize(const ResizeEvent& event)
 {
-  mScreenWidth = event.newSize.x;
-  mScreenHeight = event.newSize.y;
+  mScreenWidth = static_cast<int>(event.newSize.x);
+  mScreenHeight = static_cast<int>(event.newSize.y);
 
   for (auto cb : mResizeCallbacks)
   {
